requirements.txt
pillow~=11.1.0
Flask~=3.1.0
rembg~=2.0.62
onnxruntime~=1.20.1
waitress~=3.0.2
numpy~=2.1.3
opencv-python~=4.11.0.86